+++ Runtime.Tests/MockSchema.cs
@@ -1,0 +1,10 @@
+﻿namespace At.Ac.FhStp.ViewJson
+{
+    internal class MockSchema : IViewJsonSchema
+    {
+
+        public static MockSchema MakeDefault() =>
+            new MockSchema();
+
+    }
+}