+++ Runtime.Tests/MockStyle.cs
@@ -1,0 +1,10 @@
+﻿namespace At.Ac.FhStp.ViewJson
+{
+    internal class MockStyle : IViewJsonStyle
+    {
+
+        public static MockStyle MakeDefault() =>
+            new MockStyle();
+
+    }
+}